% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glades.env}
\alias{glades.env}
\title{glades.env title}
\description{
glades.env description
}
\references{
Baker, ME and RS King.  2010. A new method for
  detecting and interpreting biodiversity and ecological
  community thresholds.  Methods in Ecology and Evolution 1(1):
  25:37.
}
\author{
M. Baker and R. King
}
