% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_ridges.R
\name{taxa_ridges}
\alias{taxa_ridges}
\title{Taxon change point ridge plots}
\usage{
taxa_ridges(titan.out, z1 = TRUE, z2 = TRUE, prob95 = FALSE,
  z.med = FALSE, xlabel = "Environmental Gradient", all = FALSE,
  xlim, ...)
}
\arguments{
\item{titan.out}{A TITAN output object.}

\item{z1}{A logical specifying whether decreasing taxa should be plotted.}

\item{z2}{A logical specifying whether decreasing taxa should be plotted.}

\item{prob95}{A logical specifying whether change-point locations should be
plotted on the basis of their 5th (for increasers) and 95th (for
decreasers) quantile versus their observed values.}

\item{z.med}{A logical specifying whether (1) change point magnitudes should
be obtained from the median z score across bootstrap replicates and (2)
whether the locations should be plotted on the basis of the 50th quantile
of change-point locations (i.e., if prob95=FALSE).}

\item{xlabel}{A character string for the x axis label.}

\item{all}{A logical specifying whether all taxa with p<0.05 should be
plotted.}

\item{xlim}{X axis limits.}

\item{...}{...}
}
\value{
A plot of decreasing and/or increasing taxon-specific change points
  along the environmental gradient.
}
\description{
Taxon change point ridge plots
}
\note{
Should not be used with output objects from TITAN v1.0.
}
\examples{

data(glades.titan)

plotTaxa(glades.titan, xlabel = "Surface Water TP (ug/l)")
ggplotTaxa(glades.titan, xlabel = "Surface Water TP (ug/l)")
taxa_ridges(glades.titan, xlabel = "Surface Water TP (ug/l)")
taxa_ridges(glades.titan, xlabel = "Surface Water TP (ug/l)", xlim = c(0, 125))

}
\references{
Baker, ME and RS King.  2010. A new method for detecting and
  interpreting biodiversity and ecological community thresholds.  Methods in
  Ecology and Evolution 1(1): 25:37.

King, RS and ME Baker  2010. Considerations for identifying and
  interpreting ecological community thresholds. Journal of the North American
  Benthological Association 29(3):998-1008.
}
\seealso{
\code{\link{plotSumz}}, \code{\link{plotCPs}}
}
\author{
M. Baker and R. King
}
\keyword{TITAN}
\keyword{kwd2}
