% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TITAN2.R
\docType{package}
\name{TITAN2}
\alias{TITAN2}
\alias{package-TITAN2}
\title{Threshold Indicator Taxa Analysis}
\description{
Uses indicator species scores across binary partitions of a sample set to
detect congruence in taxon-specific changes of abundance and occurrence
frequency along an environmental gradient as evidence of an ecological
community threshold.
}
\details{
Relevant references include Baker and King (2010)
\url{doi:10.1111/j.2041-210X.2009.00007.x}, King and Baker (2010)
\url{doi:10.1899/09-144.1}, and Baker and King (2013) \url{doi:10.1899/12-142.1}.
}
