% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sumz_density.R
\name{plot_sumz_density}
\alias{plot_sumz_density}
\title{Plot community level change}
\usage{
plot_sumz_density(titan.out, filter = TRUE, sumz = TRUE,
  points = FALSE, ribbon = TRUE, density = TRUE,
  change_points = TRUE, sumz1 = TRUE, sumz2 = TRUE,
  xmin = min(titan.out$env), xmax = max(titan.out$envcls),
  xlabel = "Environmental Gradient", y1label = NULL,
  y2label = "Density", alpha1 = 0.65, alpha2 = 0.5,
  col1 = "steelblue4", col2 = "red", axis_lab_size = 20,
  legend.position = c(0.9, 0.9), ...)
}
\arguments{
\item{titan.out}{titan.out}

\item{filter}{filter}

\item{sumz}{sumz}

\item{points}{points}

\item{ribbon}{ribbon}

\item{density}{density}

\item{change_points}{change_points}

\item{sumz1}{sumz1}

\item{sumz2}{sumz2}

\item{xmin}{xmin}

\item{xmax}{xmax}

\item{xlabel}{xlabel}

\item{y1label}{y1label}

\item{y2label}{y2label}

\item{alpha1}{alpha1}

\item{alpha2}{alpha2}

\item{col1}{col1}

\item{col2}{col2}

\item{axis_lab_size}{axis_lab_size}

\item{legend.position}{legend.position}

\item{...}{...}
}
\value{
A plot
}
\description{
Plot community level change
}
\note{
Should not be used with output objects from TITAN v1.0.
}
\examples{

data(glades.titan)

plot_sumz_density(glades.titan)



}
\references{
Baker, ME and RS King.  2010. A new method for detecting and
  interpreting biodiversity and ecological community thresholds.  Methods in
  Ecology and Evolution 1(1): 25:37.

King, RS and ME Baker  2010. Considerations for identifying and
  interpreting ecological community thresholds. Journal of the North American
  Benthological Association 29(3):998-1008.
}
\seealso{
\code{\link{plotSumz}}, \code{\link{plotCPs}}
}
\author{
M. Baker and R. King
}
\keyword{TITAN}
\keyword{kwd2}
